requirements.txt
gunicorn==20.1.0
Flask==2.3.2
python-binance==1.0.17
binance-futures-connector==3.3.1
